Integrate record insertion in DB into pipeline
